Add setup.py, rename project to from Hasher -> 'hashes'
This is the main event. This commit turns Hasher into a packaged Distribution,
suitable for uploading to PyPI, and able to be installed with pip.

Because the name Hasher is already taken I've taken the liberty of choosing
a new name: hashes. The current owner of https://pypi.python.org/pypi/hashes
is happy to transfer it.