Remove dynamic importing of hash modules
The existing code relied on the cwd being inside a checkout.
It also assumed too much about Python module layout
and whether the packing is installed unpacked on a filesystems.

In the interest of simplicity I've used static imports for the
first release. We can add a proper plugin system later.